Factor out common categories code
