+++ src/pages/BasePersonalisationPage.js
@@ -6,15 +6,14 @@ import React from 'react';
 import Router from "react-router";
 import mui from "material-ui";
 import reactMixin from "react-mixin";
-import _ from "underscore";
 
-import categories from '../constants/categories';
+import BaseCategoriesPage from './BaseCategoriesPage';
 import components from '../components';
 import icons from '../icons';
 
 /*::`*/@reactMixin.decorate(Router.Navigation)/*::`;*/
 /*::`*/@reactMixin.decorate(Router.State)/*::`;*/
-class BasePersonalisationPage extends React.Component {
+class BasePersonalisationPage extends BaseCategoriesPage {
     constructor(props: Object) {
         super(props);
         this.state = {
@@ -31,24 +30,6 @@ class BasePersonalisationPage extends React.Component {
         return {
             controller: this,
         };
-    }
-
-    // flow:disable not supported yet
-    get category(): categories.Category {
-        if (this._category) {
-            return this._category;
-        }
-
-        var category = _.findWhere(categories, {
-            key: this.props.params.page,
-        });
-
-        if (category === undefined) {
-            throw "No such category " + this.props.params.page;
-        }
-
-        this._category = category;
-        return category;
     }
 
     previousStep(): void {
